Extract `Account` into `src/Account.js`
